+++ ThunderED/Json/SystemName.cs
@@ -6,18 +6,18 @@ namespace ThunderED.Json
     {
         public class SystemName
         {
-            public int constellation_id { get; set; }
+            public long constellation_id { get; set; }
             public string name { get; set; }
             public Planet[] planets { get; set; }
             public Position position { get; set; }
             public string security_class;
             public float security_status { get; set; }
-            public int star_id { get; set; }
-            public int[] stargates { get; set; }
-            public int system_id { get; set; }
+            public long star_id { get; set; }
+            public long[] stargates { get; set; }
+            public long system_id { get; set; }
 
             [JsonIgnore]
-            public int? DB_RegionId;
+            public long? DB_RegionId;
         }
     }
 }
